+++ samples/ControlGallery/ControlGallery/Pages/ClipPage.xaml
@@ -12,16 +12,21 @@
                    FontAttributes="Bold"
                    HorizontalOptions="Center" />
 
+            <!-- Ellipse Clip -->
             <Border Stroke="Gray"
                     Padding="12">
                 <Border.StrokeShape>
                     <RoundRectangle CornerRadius="12" />
                 </Border.StrokeShape>
-                <VerticalStackLayout Spacing="12">
+                <VerticalStackLayout Spacing="12"
+                                     WidthRequest="320"
+                                     HorizontalOptions="Center">
                     <Label Text="Ellipse Clip"
                            FontAttributes="Bold" />
                     <Grid HeightRequest="160"
-                          WidthRequest="320">
+                          WidthRequest="320"
+                          HorizontalOptions="Center"
+                          VerticalOptions="Center">
                         <Grid.Background>
                             <LinearGradientBrush StartPoint="0,0"
                                                  EndPoint="1,1">
@@ -45,16 +50,21 @@
                 </VerticalStackLayout>
             </Border>
 
+            <!-- Rounded Rectangle Clip -->
             <Border Stroke="Gray"
                     Padding="12">
                 <Border.StrokeShape>
                     <RoundRectangle CornerRadius="12" />
                 </Border.StrokeShape>
-                <VerticalStackLayout Spacing="12">
+                <VerticalStackLayout Spacing="12"
+                                     WidthRequest="320"
+                                     HorizontalOptions="Center">
                     <Label Text="Rounded Rectangle Clip"
                            FontAttributes="Bold" />
                     <Grid HeightRequest="140"
                           WidthRequest="320"
+                          HorizontalOptions="Center"
+                          VerticalOptions="Center"
                           Padding="12">
                         <Grid.Background>
                             <LinearGradientBrush StartPoint="0,0"
@@ -82,16 +92,21 @@
                 </VerticalStackLayout>
             </Border>
 
+            <!-- Custom Path Clip -->
             <Border Stroke="Gray"
                     Padding="12">
                 <Border.StrokeShape>
                     <RoundRectangle CornerRadius="12" />
                 </Border.StrokeShape>
-                <VerticalStackLayout Spacing="12">
+                <VerticalStackLayout Spacing="12" 
+                                     WidthRequest="320" 
+                                     HorizontalOptions="Center">
                     <Label Text="Custom Path Clip"
                            FontAttributes="Bold" />
                     <Grid HeightRequest="150"
-                          WidthRequest="320">
+                          WidthRequest="320"
+                          HorizontalOptions="Center"
+                          VerticalOptions="Center">
                         <Grid.Background>
                             <LinearGradientBrush StartPoint="0,0"
                                                  EndPoint="1,0">
