+++ samples/ControlGallery/ControlGallery/Pages/ShadowPage.xaml
@@ -12,6 +12,7 @@
                    FontAttributes="Bold"
                    HorizontalOptions="Center" />
 
+            <!-- Soft Card Shadow -->
             <Border Stroke="Gray"
                     Padding="12">
                 <Border.StrokeShape>
@@ -22,7 +23,8 @@
                            FontAttributes="Bold" />
                     <Border HeightRequest="140"
                             BackgroundColor="#F3F4F6"
-                            Padding="16">
+                            Padding="12"
+                            Margin="24">
                         <Border.Shadow>
                             <Shadow Brush="#2563EB"
                                     Offset="0,12"
@@ -39,6 +41,7 @@
                 </VerticalStackLayout>
             </Border>
 
+            <!-- Directional Shadow -->
             <Border Stroke="Gray"
                     Padding="12">
                 <Border.StrokeShape>
@@ -49,7 +52,8 @@
                            FontAttributes="Bold" />
                     <Border HeightRequest="140"
                             BackgroundColor="#FEF3C7"
-                            Padding="16">
+                            Padding="12"
+                            Margin="24">
                         <Border.Shadow>
                             <Shadow Brush="#EA580C"
                                     Offset="18,6"
@@ -66,6 +70,7 @@
                 </VerticalStackLayout>
             </Border>
 
+            <!-- Raised Card -->
             <Border Stroke="Gray"
                     Padding="12">
                 <Border.StrokeShape>
@@ -76,7 +81,8 @@
                            FontAttributes="Bold" />
                     <Border HeightRequest="140"
                             BackgroundColor="#ECFEFF"
-                            Padding="16">
+                            Padding="12"
+                            Margin="24">
                         <Border.Shadow>
                             <Shadow Brush="#0EA5E9"
                                     Offset="0,20"
