+++ src/Avalonia.Controls.Maui/Handlers/SwipeViewHandler.cs
@@ -1,5 +1,6 @@
 using Microsoft.Maui;
 using Microsoft.Maui.Handlers;
+using Avalonia.Controls.Maui.Extensions;
 using PlatformView = Avalonia.Controls.Maui.Swipe;
 
 namespace Avalonia.Controls.Maui.Handlers;
