+++ src/Avalonia.Controls.Maui/Handlers/ViewHandlerOfT.cs
@@ -39,10 +39,27 @@ public abstract partial class ViewHandler<TVirtualView, TPlatformView> : ViewHan
     public sealed override void SetVirtualView(IElement view) =>
         SetVirtualView((IView)view);
 
+    /// <summary>
+    /// Gets or sets a factory function for creating platform views.
+    /// </summary>
+    /// <remarks>This property allows customization of platform view creation by providing a factory function
+    /// that takes a handler instance and returns a platform view. If not set, the default <see cref="CreatePlatformView"/> method is used.</remarks>
     public static Func<ViewHandler<TVirtualView, TPlatformView>, TPlatformView>? PlatformViewFactory { get; set; }
 
+    /// <summary>
+    /// Creates a new platform view instance.
+    /// </summary>
+    /// <returns>A new instance of the platform-specific view.</returns>
+    /// <remarks>This method is called when the handler needs to create a platform view for the virtual view.
+    /// Derived classes must implement this method to provide the appropriate platform view creation logic.</remarks>
     protected abstract TPlatformView CreatePlatformView();
 
+    /// <summary>
+    /// Connects the handler to the platform view.
+    /// </summary>
+    /// <param name="platformView">The platform view to connect to.</param>
+    /// <remarks>This method is called when the handler is connected to its platform view.
+    /// It initializes gesture manager for controls views and can be overridden to perform additional connection logic.</remarks>
     protected virtual void ConnectHandler(TPlatformView platformView)
     {
         // Initialize gesture manager for IControlsView
@@ -52,6 +69,12 @@ public abstract partial class ViewHandler<TVirtualView, TPlatformView> : ViewHan
         }
     }
 
+    /// <summary>
+    /// Disconnects handler from platform view.
+    /// </summary>
+    /// <param name="platformView">The platform view to disconnect from.</param>
+    /// <remarks>This method is called when handler is disconnected from its platform view.
+    /// It disposes the gesture manager and can be overridden to perform additional cleanup logic.</remarks>
     protected virtual void DisconnectHandler(TPlatformView platformView)
     {
         _gestureManager?.Dispose();
